feat: refactor reconciliation logic to separate concerns, fixes RHOAIENG-25536 (#251)
* feat: refactor reconciliation logic to separate concerns, fixes RHOAIENG-25536


* fix: move getRouteLabels to istio functions


* fix: minor code cleanup


---------
